+++ symengine/CMakeLists.txt
@@ -21,7 +21,8 @@ set(SRC
     pow.cpp
     functions.cpp
     monomials.cpp
-    polynomial.cpp
+    uintpoly.cpp
+    uexprpoly.cpp
     series.cpp
     series_generic.cpp
     rings.cpp
@@ -68,14 +69,14 @@ endif()
 
 # Needed for "make install"
 set(HEADERS
-    add.h        symengine_assert.h  functions.h  monomials.h  polynomial.h polynomial_multivariate.h number.h    rings.h
+    add.h        symengine_assert.h  functions.h  monomials.h  uintpoly.h uexprpoly.h polynomial_multivariate.h number.h    rings.h
     basic.h      symengine_rcp.h     integer.h    mul.h        pow.h       constants.h
     symbol.h     expression.h        parser.h     mp_class.h   mp_wrapper.h
     basic-inl.h  dict.h           matrix.h     ntheory.h    rational.h complex.h
     visitor.h    eval_double.h    diophantine.h cwrapper.h  printer.h  real_double.h
     eval_mpfr.h  eval_arb.h       eval_mpc.h     complex_double.h         series_visitor.h
     real_mpfr.h  complex_mpc.h    type_codes.inc lambda_double.h series.h series_piranha.h
-    basic-methods.inc   series_flint.h  series_generic.h sets.h  derivative.h   subs.h  uint_base.h
+    basic-methods.inc   series_flint.h  series_generic.h sets.h  derivative.h   subs.h  upolybase.h
 )
 
 # Configure SymEngine using our CMake options:
