+++ symengine/tests/basic/test_polynomial_multivariate.cpp
@@ -3,7 +3,8 @@
 #include <iostream>
 
 #include <symengine/mul.h>
-#include <symengine/polynomial.h>
+#include <symengine/uintpoly.h>
+#include <symengine/uexprpoly.h>
 #include <symengine/polynomial_multivariate.h>
 #include <symengine/pow.h>
 #include <symengine/printer.h>
